+++ pydigitalstrom/devices/scene.py
@@ -4,22 +4,50 @@ from pydigitalstrom.devices.base import DSDevice
 
 
 class DSScene(DSDevice):
-    URL_TURN_ON = '/json/zone/callScene?id={zone_id}&sceneNumber={scene_id}'
-    URL_TURN_OFF = '/json/zone/undoScene?id={zone_id}&sceneNumber={scene_id}'
+    URL_TURN_ON = '/json/zone/callScene?id={zone_id}&' \
+                  'sceneNumber={scene_id}&force=true'
 
-    def __init__(self, client: DSClient, data, *args, **kwargs):
-        data['id'] = '{zone_id}.{scene_id}'.format(
-            zone_id=data['zone_id'], scene_id=data['scene_id'])
-        data['name'] = '{zone} / {name}'.format(
-            zone=data['zone_name'], name=data['scene_name'])
-        super().__init__(client=client, data=data, *args, **kwargs)
+    def __init__(self, client: DSClient, zone_id, zone_name, scene_id,
+                 scene_name, *args, **kwargs):
+        self.zone_id = zone_id
+        self.zone_name = zone_name
+        self.scene_id = scene_id
+        self.scene_name = scene_name
+
+        device_id = '{zone_id}.{scene_id}'.format(
+            zone_id=self.zone_id, scene_id=self.scene_id)
+        device_name = '{zone} / {name}'.format(
+            zone=self.zone_name, name=self.scene_name)
+
+        super().__init__(client=client, device_id=device_id,
+                         device_name=device_name, *args, **kwargs)
 
     async def turn_on(self):
         await self.request(url=self.URL_TURN_ON.format(
-            zone_id=self._data['zone_id'], scene_id=self._data['scene_id']),
-            check_result=False)
+            zone_id=self.zone_id, scene_id=self.scene_id), check_result=False)
 
-    async def turn_off(self):
-        await self.request(url=self.URL_TURN_OFF.format(
-            zone_id=self._data['zone_id'], scene_id=self._data['scene_id']),
+
+class DSColorScene(DSDevice):
+    URL_TURN_ON = '/json/zone/callScene?id={zone_id}&' \
+                  'sceneNumber={scene_id}&groupID={color}&force=true'
+
+    def __init__(self, client: DSClient, zone_id, zone_name, scene_id,
+                 scene_name, color, *args, **kwargs):
+        self.zone_id = zone_id
+        self.zone_name = zone_name
+        self.scene_id = scene_id
+        self.scene_name = scene_name
+        self.color = color
+
+        device_id = '{zone_id}.{color}.{scene_id}'.format(
+            zone_id=self.zone_id, color=self.color, scene_id=self.scene_id)
+        device_name = '{zone} / {name}'.format(
+            zone=self.zone_name, name=self.scene_name)
+
+        super().__init__(client=client, device_id=device_id,
+                         device_name=device_name, *args, **kwargs)
+
+    async def turn_on(self):
+        await self.request(url=self.URL_TURN_ON.format(
+            zone_id=self.zone_id, color=self.color, scene_id=self.scene_id),
             check_result=False)