+++ pydigitalstrom/exceptions.py
@@ -9,13 +9,5 @@ class DSRequestException(DSException):
     pass
 
 
-class DSUnsupportedException(DSException):
-    pass
-
-
 class DSCommandFailedException(DSException):
-    pass
-
-
-class DSParameterException(DSException):
     pass