+++ tests/devices/test_scene.py
@@ -2,41 +2,53 @@
 import aiounittest
 from unittest.mock import Mock, patch
 
-from pydigitalstrom.devices.scene import DSScene
+from pydigitalstrom.devices.scene import DSScene, DSColorScene
 from tests.common import get_testclient
 
 
 class TestScene(aiounittest.AsyncTestCase):
     def test_initial(self):
         device = DSScene(
-            client=get_testclient(), data=dict(zone_id=1, zone_name='zone',
-                                               scene_id=2, scene_name='scene'))
-        self.assertEqual(device._data, dict(zone_id=1, zone_name='zone',
-                                            scene_id=2, scene_name='scene',
-                                            id='1.2', name='zone / scene'))
+            client=get_testclient(), zone_id=1, zone_name='zone',
+            scene_id=2, scene_name='scene')
+        self.assertEqual(device.zone_id, 1)
+        self.assertEqual(device.zone_name, 'zone')
+        self.assertEqual(device.scene_id, 2)
+        self.assertEqual(device.scene_name, 'scene')
 
     async def test_turn_on(self):
         with patch('pydigitalstrom.devices.scene.DSScene.request',
                    Mock(return_value=aiounittest.futurized(dict()))) as \
                 mock_request:
             device = DSScene(
-                client=get_testclient(), data=dict(
-                    zone_id=1, zone_name='zone',
-                    scene_id=2, scene_name='scene'))
+                client=get_testclient(), zone_id=1, zone_name='zone',
+                scene_id=2, scene_name='scene')
             await device.turn_on()
             mock_request.assert_called_with(
-                url='/json/zone/callScene?id=1&sceneNumber=2',
+                url='/json/zone/callScene?id=1&sceneNumber=2&force=true',
                 check_result=False)
 
-    async def test_turn_off(self):
-        with patch('pydigitalstrom.devices.scene.DSScene.request',
+
+class TestColorScene(aiounittest.AsyncTestCase):
+    def test_initial(self):
+        device = DSColorScene(
+            client=get_testclient(), zone_id=1, zone_name='zone',
+            scene_id=2, scene_name='scene', color=1)
+        self.assertEqual(device.zone_id, 1)
+        self.assertEqual(device.zone_name, 'zone')
+        self.assertEqual(device.scene_id, 2)
+        self.assertEqual(device.scene_name, 'scene')
+        self.assertEqual(device.color, 1)
+
+    async def test_turn_on(self):
+        with patch('pydigitalstrom.devices.scene.DSColorScene.request',
                    Mock(return_value=aiounittest.futurized(dict()))) as \
                 mock_request:
-            device = DSScene(
-                client=get_testclient(), data=dict(
-                    zone_id=1, zone_name='zone',
-                    scene_id=2, scene_name='scene'))
-            await device.turn_off()
+            device = DSColorScene(
+                client=get_testclient(), zone_id=1, zone_name='zone',
+                scene_id=2, scene_name='scene', color=1)
+            await device.turn_on()
             mock_request.assert_called_with(
-                url='/json/zone/undoScene?id=1&sceneNumber=2',
+                url='/json/zone/callScene?id=1&sceneNumber=2&groupID=1&'
+                    'force=true',
                 check_result=False)